+++ app/pages/module.js
@@ -121,7 +121,7 @@ const pageContent = [
             className: 'table-content-filter',
             data: `
                 #.currentProfile.callFramesTimingsFiltered.entries.[entry.module = @]
-                    .zip(=> entry, #.currentProfile.codesByCallFrame, => function)
+                    .zip(=> entry, #.currentProfile.codesByCallFrame, => callFrame)
                     .({
                         $entry: left.entry;
 
@@ -149,6 +149,14 @@ const pageContent = [
                     `,
                     cols: [
                         ...timingCols,
+                        {
+                            header: '',
+                            colWhen: '$[=>right]',
+                            sorting: 'right.hotness | $ = "hot" ? 3 : $ = "warm" ? 2 : $ = "cold" ? 1 : 0 desc',
+                            data: 'right',
+                            contentWhen: 'hotness = "hot" or hotness = "warm"',
+                            content: 'hotness-icon{ hotness, topTier }'
+                        },
                         { header: 'Call frame',
                             className: 'subject-name',
                             sorting: 'name ascN',